Standard installation of the extension
1. Enable developer mode
on the chrome tab://extensions/
On the Extensions page, toggle Developer Mode
For ON in the upper right corner.
2. Install the extension manually
You can install extensions directly from the zip file. 
Click on Download the unpacked version to download the extension from your computer.
Find the extensions folder (unpacked extension).
Click on the folder to install the extension manually.
3. Everything is ready, you have an extension!Pin it as usual in the row on the top right and use it to your heart's content.
Be careful, I am a developer from Russia and the entire interface is in Russian, but it does not matter how the extension works, the main thing is when you install the extension, activate it with the first button. The second button is responsible for resetting the reward counter.